Fix up based on feedback
