+++ app/SharedWebComponents/Components/Answer.razor.cs
@@ -1,7 +1,5 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 
-using System.Reflection.Metadata;
-
 namespace SharedWebComponents.Components;
 
 public sealed partial class Answer
