+++ app/SharedWebComponents/Pages/VoiceChat.razor.cs
@@ -89,8 +89,12 @@ public sealed partial class VoiceChat : IDisposable
                         }
                         catch
                         {
-                            // TODO: try and not use the method above as it is not really useful
-                            //       since the implementation checks for a specific type
+                            // TODO: Find a better way to do this
+                            // The Blazor.SpeechSynthesis.WebAssembly API supports the callback,
+                            // however the underlying code does not do this using a DI-friendly way.
+                            // The code assumes the concrete implementation for the ISpeechSynthesisService
+                            // service is the concrete Web Assembly type which is not valid. However, the
+                            // alternate API is something MAUI can implement.
 
                             SpeechSynthesis.Speak(utterance);
